I have added a dictionary.txt file with the 10 000 most common words according to MIT.
I use HashSet for this list but changed the commonPasswords back to an ArrayList. The reason is that I want to tell the user what place their password ranks amongst common passwords. A HashSet does not preserve the order they are stored in inside the txt file. But since I do not care what order the word comes in inside the dictionary I made this a HashSet instead.

I also discovered a small mistake in generateCombinations where the count only worked until it reached z, and then it would reset.

I also changed the printCombinations as I found that it wouldn't find a password from the txt files if that password was smaller than the given startLength or longer than maxLength. This has now been fixed so that it will go through all the txt files regardless of start and max length. It also didn't stop properly when it found the password in a txt file and started generating passwords, so I added the System.exit(0); lines in both the common and dictionary password loops.

It seems like everything works now and the next step will probably be to make it more efficient by combining words or numbers.